Added file lock for multiprocessing
